Implement graphing mode on Tab 4
- Repurposed Tab 4 (F4) for function graphing.
- Added `ui_draw_graph()` to plot mathematical expressions using `tinyexpr` (supporting variable `x`).
- Updated `ui_redraw_tab_content()` to render a coordinate system and the plotted function on the 4th tab.
- Integrated `spi_draw_pixel()` with 32-bit color support in `lcdspi`.
- Added explicit null-termination for history expressions in `ui_add_to_history()`.
- Enhanced `handle_keyboard()` in `main.c` to support entering expressions for graphing.

+++ lcdspi/lcdspi.h
@@ -113,7 +113,7 @@ extern void spi_write_command(unsigned char data);
 extern void spi_write_cd(unsigned char command, int data, ...);
 extern void spi_write_data24(uint32_t data);
 
-extern void spi_draw_pixel(uint16_t x, uint16_t y, uint16_t color) ;
+extern void spi_draw_pixel(uint16_t x, uint16_t y, uint32_t color) ;
 extern void draw_rect_spi(int x1, int y1, int x2, int y2, int c) ;
 extern void lcd_putc(uint8_t devn, uint8_t c);
 extern int  lcd_getc(uint8_t devn);
